Refactored transaction routes to split routes into their respective files.
